test(http): Extract memory tests from `ApplicationTest` to `ApplicationMemoryTest` class.
